Refactor functions related to creation of txt file under new Package
